Fix sorting of the signature params. The current sorting has problems with long responses and indexed fields.
According to the Buckaroo documentation, the sort algorithm uses the following order:
symbols first, then numbers, then case insensitive letters.

This is implemented by defining the char order as a constant. Currently,
the only symbol defined is '_', because this seems to be the only symbol that
it used in field names.